move key types to key
