+++ extension/chrome/elements/composer/composer-my-pubkey.ts
@@ -7,7 +7,7 @@ import { Ui } from '../../../js/common/browser/ui.js';
 import { Catch } from '../../../js/common/platform/catch.js';
 import { Lang } from '../../../js/common/lang.js';
 import { Store } from '../../../js/common/platform/store.js';
-import { KeyInfo } from '../../../js/common/core/pgp-msg.js';
+import { KeyInfo } from '../../../js/common/core/pgp-key.js';
 import { ApiErr } from '../../../js/common/api/error/api-error.js';
 import { PgpKey } from '../../../js/common/core/pgp-key.js';
