+++ extension/chrome/elements/composer/composer-recipients.ts
@@ -5,7 +5,6 @@
 import { Composer } from './composer.js';
 import { Str, Value } from '../../../js/common/core/common.js';
 import { ProviderContactsQuery, Recipients } from '../../../js/common/api/email_provider/email_provider_api.js';
-import { Contact } from '../../../js/common/core/pgp-msg.js';
 import { Xss } from '../../../js/common/platform/xss.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { Google } from '../../../js/common/api/google.js';
@@ -20,7 +19,7 @@ import { RecipientType, ChunkedCb } from '../../../js/common/api/api.js';
 import { Store } from '../../../js/common/platform/store.js';
 import { PUBKEY_LOOKUP_RESULT_FAIL, PUBKEY_LOOKUP_RESULT_WRONG } from './composer-errs.js';
 import { ApiErr } from '../../../js/common/api/error/api-error.js';
-import { PgpKey } from '../../../js/common/core/pgp-key.js';
+import { PgpKey, Contact } from '../../../js/common/core/pgp-key.js';
 
 export class ComposerRecipients extends ComposerComponent {
   private addedRecipients: RecipientElement[] = [];
