+++ extension/js/common/platform/store.ts
@@ -4,7 +4,7 @@
 
 import { Value, Str, Dict } from '../core/common.js';
 import { mnemonic } from '../core/mnemonic.js';
-import { KeyInfo, Contact } from '../core/pgp-msg.js';
+import { KeyInfo, Contact } from '../core/pgp-key.js';
 import { SubscriptionInfo, PaymentMethod, ProductLevel, FcUuidAuth } from '../api/backend.js';
 import { BrowserMsg, BgNotReadyErr } from '../browser/browser-msg.js';
 import { Catch, UnreportableError } from './catch.js';
